# server/openapi.yaml
openapi: 3.0.0
info:
  title: Glass Fence API Reference
  version: "3.0.0"
  description: |
    Glass Fence uses WebRTC and WebSocket besides REST API to provide real-time video streaming, interactive collaboration, and remote desktop functionality.

    ### Core Technologies

    - **Video Streaming over WebRTC**
      - Glass Fence leverages **WebRTC** for real-time, low-latency, high-quality video and audio streaming.
      - Keyboard and mouse events are transmitted over the same WebRTC connection for seamless interaction.

    - **Real-Time Signaling and Communication over WebSocket**
      - **WebSocket** is used for sSignaling in WebRTC (e.g., session establishment, ICE candidate exchange).  
      - Live chat, session updates, and keyboard/mouse fallback are handled over WebSocket.

    - **REST API for Everything Else**
      - **Authentication & Session Management** - Secure user authentication and session handling.  
      - **Room and User Management** - Creating, modifying, and controlling user access to rooms.  
      - **Desktop Controls** - Adjusting screen configurations, managing keyboard and mouse events.
      - **System Health & Metrics** - Retrieving system health checks and operational insights.  

    ### How It All Works Together

    1. **WebRTC** handles media streaming for **real-time video, audio and data**.
    2. **WebSocket** ensures a **persistent, bidirectional** connection for low-latency interactions.  
...

...
  license:
    name: Apache 2.0
...
servers:
...
paths:
...
  /api/room/upload/drop:
    post:
      tags:
        - room-upload
      summary: Upload and Drop File
      description: Upload a file and drop it at a specified location.
      operationId: uploadDrop
      responses:
        '204':
          description: File uploaded and dropped successfully.
        '400':
          description: Unable to upload file.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          description: Unable to process uploaded file.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                x:
                  type: number
                  description: X coordinate of drop.
                y:
                  type: number
                  description: Y coordinate of drop.
                files:
                  type: array
                  description: Files to be uploaded.
                  items:
                    type: string
                    format: binary
        required: true
  /api/room/upload/dialog:
    post:
      tags:
        - room-upload
      summary: Upload File to Dialog
      description: Upload a file to a dialog.
      operationId: uploadDialog
      responses:
        '204':
          description: File uploaded to dialog successfully.
        '400':
          description: Unable to upload file.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          description: No upload dialog prompt active.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
        '500':
          description: Unable to process uploaded file.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'
      requestBody:
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                files:
                  type: array
                  description: Files to be uploaded.
                  items:
                    type: string
                    format: binary
        required: true
    delete:
      tags:
        - room-upload
      summary: Close File Chooser Dialog
      description: Close the file chooser dialog.
      operationId: uploadDialogClose
      responses:
        '204':
          description: File chooser dialog closed successfully.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          description: No upload dialog prompt active.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorMessage'

  #
  # members
  #
...
components:
  securitySchemes:
    CookieAuth:
      type: apiKey
      in: cookie
      name: GF_SESSION
      description: Authentication using a session cookie.
    BearerAuth:
      type: http
      scheme: bearer
      description: Authentication using a Bearer token.
    TokenAuth:
      type: apiKey
      in: query
      name: token
      description: Authentication using a token passed as a query parameter.
  responses:
...
    Unauthorized:
      description: The request requires user authentication.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorMessage'
    Forbidden:
      description: The server understood the request, but refuses to authorize it.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorMessage'
  schemas:
    ErrorMessage:
      type: object
      properties:
        message:
          type: string
          description: Detailed error message.
...
security:
  - BearerAuth: []
  - CookieAuth: []
  - TokenAuth: []